refactor: split main.js into modules
