+++ js/app.js
@@ -1,0 +1,1 @@
+export const app = {};
